… src/System.CommandLine.ApiCompatibility.Tests/ApiCompatibilityApprovalTests.System_CommandLine_api_is_not_changed.approved.txt
… src/System.CommandLine/Help/HelpOption.cs
… src/System.CommandLine/Help/HelpResult.cs
… src/System.CommandLine/Help/VersionOption.cs
… src/System.CommandLine/Invocation/InvocationPipeline.cs
… src/System.CommandLine/Invocation/ParseDirectiveResult.cs
… src/System.CommandLine/Invocation/ParseErrorResult.cs
… src/System.CommandLine/Invocation/TypoCorrection.cs
… src/System.CommandLine/ParseResult.cs
… src/System.CommandLine/Parsing/ArgumentResult.cs
… src/System.CommandLine/Parsing/ParseOperation.cs
… src/System.CommandLine/Parsing/SymbolResultTree.cs