@adamsitnik 2023-02-03
Updated benchmark numbers:

|        Method |      Args | Mean before | Mean after | Allocated before | Allocated after |
|-------------- |---------- |------------:|-----------:|-----------------:|----------------:|
|  DefaultsSync | String[0] |   29.035 us |  26.464 us |         10,264 B |         7,960 B |
| DefaultsAsync | String[0] |   29.850 us |  26.742 us |         10,264 B |         7,960 B |
|   MinimalSync | String[0] |    1.316 us |   1.321 us |          3,264 B |         3,336 B |
|  MinimalAsync | String[0] |    1.448 us |   1.445 us |          3,264 B |         3,336 B |
|  DefaultsSync | String[4] |   31.213 us |  28.696 us |         11,760 B |         9,456 B |
| DefaultsAsync | String[4] |   31.244 us |  29.251 us |         11,760 B |         9,456 B |
|   MinimalSync | String[4] |    2.415 us |   2.365 us |          4,760 B |         4,832 B |
|  MinimalAsync | String[4] |    2.497 us |   2.535 us |          4,760 B |         4,832 B |